Update index.php
The modification is backwards compatible with existing version, but it adds the ability to enhance the buttons with classes ($className) such as btn-danger, btn-warning, btn-success, etc.  The $otherAtt provides a means of adding attributes such as styles="width:50%" and/or autofocus.  These modifications also require a modification to the userbuttons file.